Renaming, standardized error handling from API calls, moving code
