+++ src/api/error.js
@@ -1,0 +1,10 @@
+// @flow
+
+const handleError = async ( e: Object ) => {
+  if ( !e.response ) { throw e; }
+  const errorJson = await e.response.json( );
+  const error = new Error( errorJson.status );
+  throw error;
+};
+
+export default handleError;
